feat: go to map in map rotation
